+++ services/routine/internal/app/image/controller.go
