+++ services/routine/internal/app/image/controller_funcs.go
